+++ package.json
@@ -26,7 +26,6 @@
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
-    "react-spring": "^9.4.4",
     "typescript": "^4.6.2"
   },
   "scripts": {
