+++ src/common/components/App/App.module.scss
@@ -12,7 +12,6 @@
   display: flex;
   flex-wrap: wrap;
   flex-direction: row-reverse;
-  margin: 0 auto;
 }
 
 .appWrapperContent {
