+++ src/common/components/BarInstallApps/BarInstallApps.module.scss
@@ -1,0 +1,22 @@
+.installAppsWrapper {
+  flex-direction: column;
+  align-items: center;
+  width: 350px;
+  font-size: 14px;
+  margin-top: 15px;
+  display: flex;
+}
+
+.barInstallApps {
+  gap: 10px;
+  margin-top: 15px;
+  display: flex;
+
+  :hover {
+    cursor: pointer;
+  }
+
+  img {
+    width: 130px;
+  }
+}
