+++ src/common/components/BarInstallApps/index.ts
@@ -1,0 +1,1 @@
+export { BarInstallApps } from "./BarInstallApps";
