+++ src/common/components/Registration/Registration.module.scss
@@ -1,0 +1,3 @@
+.registrationWrapper{
+  margin: 100px auto;
+}
